clippy

+++ sparse_strips/vello_common/src/glyph.rs
@@ -435,7 +435,7 @@ fn prepare_colr_glyph<'a>(
     // exactly one pixel (or more) per device pixel, to ensure that no quality is lost. Therefore,
     // we simply use the scaling/skewing factor to calculate how much to scale each axis by.
     let (scale_factor_x, scale_factor_y) = {
-        let (x_vec, y_vec) = x_y_advances(&transform.pre_scale(f64::from(font_size_scale)));
+        let (x_vec, y_vec) = x_y_advances(&transform.pre_scale(font_size_scale));
         (x_vec.length(), y_vec.length())
     };
